Listen for keydown in the stepper
Listening for keydown instead of keyup allows the user to hold down an
arrow key to continue stepping. This also fixes the bug that caused both
the stepper value and the count to increase in response to the same
keypress because now the main application and the stepper are listening
for the same event.